requirements.txt
Django==4.2.1
pillow==9.5.0
djangorestframework==3.14.0